+++ tgui/packages/tgui/interfaces/Newscaster/constants.ts
@@ -1,0 +1,8 @@
+export const NEWSCASTER_SCREEN_MAIN = 'Main Menu';
+export const NEWSCASTER_SCREEN_NEWCHANNEL = 'New Channel';
+export const NEWSCASTER_SCREEN_VIEWLIST = 'View List';
+export const NEWSCASTER_SCREEN_NEWSTORY = 'New Story';
+export const NEWSCASTER_SCREEN_PRINT = 'Print';
+export const NEWSCASTER_SCREEN_NEWWANTED = 'New Wanted';
+export const NEWSCASTER_SCREEN_VIEWWANTED = 'View Wanted';
+export const NEWSCASTER_SCREEN_SELECTEDCHANNEL = 'View Selected Channel';
