+++ tgui/packages/tgui/interfaces/OperatingComputer/types.ts
@@ -1,0 +1,38 @@
+import { BooleanLike } from 'common/react';
+
+export type Data = {
+  hasOccupant: BooleanLike;
+  occupant: occupant;
+  verbose: BooleanLike;
+  oxyAlarm: number;
+  choice: BooleanLike;
+  health: BooleanLike;
+  crit: BooleanLike;
+  healthAlarm: BooleanLike;
+  oxy: BooleanLike;
+};
+
+export type occupant = {
+  name: string;
+  stat: number;
+  health: number;
+  maxHealth: number;
+  minHealth: number;
+  bruteLoss: number;
+  oxyLoss: number;
+  toxLoss: number;
+  fireLoss: number;
+  paralysis: number;
+  hasBlood: BooleanLike;
+  bodyTemperature: number;
+  maxTemp: number;
+  temperatureSuitability: number;
+  btCelsius: number;
+  btFaren: number;
+  pulse: number;
+  bloodLevel: number;
+  bloodMax: number;
+  bloodPercent: number;
+  bloodType: string;
+  surgery: { name: string; currentStage: string; nextSteps: string[] }[] | null;
+};
